# docs/openapi/mescopilot.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MesCopilot.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/device-connections": {
      "get": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceConnectionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceConnectionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceConnectionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-connections/{id}": {
      "put": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConnectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-connections/{id}/test": {
      "post": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-connections/{id}/start": {
      "post": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-connections/{id}/stop": {
      "post": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-connections/{id}/realtime": {
      "get": {
        "tags": [
          "DeviceConnections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "Active tenant selected from the authenticated user's memberships.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentStatusUpdate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentStatusUpdate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentStatusUpdate"
                }
              }
            }
          },
          "400": {
            "description": "Tenant selection is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Tenant membership is missing or inactive.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceConnectionDto": {
        "required": [
          "allowInsecure",
          "endpoint",
          "equipmentCode",
          "equipmentId",
          "hasCredentials",
          "host",
          "id",
          "isEnabled",
          "name",
          "port",
          "protocol",
          "useTls"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/DeviceProtocol"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "endpoint": {
            "type": "string"
          },
          "hasCredentials": {
            "type": "boolean"
          },
          "useTls": {
            "type": "boolean"
          },
          "allowInsecure": {
            "type": "boolean"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "lastConnectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastErrorCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeviceConnectionRequest": {
        "required": [
          "allowInsecure",
          "endpoint",
          "equipmentId",
          "host",
          "name",
          "pollIntervalMilliseconds",
          "port",
          "protocol",
          "registerAddress",
          "registerCount",
          "unitId",
          "useTls"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "equipmentId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "protocol": {
            "$ref": "#/components/schemas/DeviceProtocol"
          },
          "host": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "endpoint": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "certificateThumbprint": {
            "type": "string",
            "nullable": true
          },
          "unitId": {
            "maximum": 247,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "registerAddress": {
            "type": "integer",
            "format": "int32"
          },
          "registerCount": {
            "maximum": 125,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "pollIntervalMilliseconds": {
            "maximum": 60000,
            "minimum": 100,
            "type": "integer",
            "format": "int32"
          },
          "useTls": {
            "type": "boolean"
          },
          "allowInsecure": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DeviceProtocol": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "EquipmentStatusUpdate": {
        "required": [
          "equipmentCode",
          "equipmentId",
          "state",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "productionLineId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "TenantProblemDetails": {
        "required": [
          "code",
          "status",
          "title"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}